kernel modules: install modules in /vsdk/demos/ path
With modules inserted in the standard path /lib/modules/<kernel
version>, it seems that /etc/init/udev would automatically insert the
VSDK modules at boot-time for Yocto roots.
This may result in an incorrect insertion order and also with overlap
with the /etc/init.d/vsdk_mod.sh script that force the VSDK module
insertion in a specific order.
The error is the following:

insmod: ERROR: could not insert module
/lib/modules/4.1.26-rt29+g8447e8c/oal_cma.ko: File exists
insmod: ERROR: could not insert module
/lib/modules/4.1.26-rt29+g8447e8c/csi.ko: File exists
[ 16.503904] CGD: ready.
insmod: ERROR: could not insert module
/lib/modules/4.1.26-rt29+g8447e8c/seq.ko: File exists
insmod: ERROR: could not insert module
/lib/modules/4.1.26-rt29+g8447e8c/fdma.ko: File exists
insmod: ERROR: could not insert module
/lib/modules/4.1.26-rt29+g8447e8c/h264enc.ko: File exists

The fix is to install the modules in the /vsdk/demos path
A default path is defined in layer.conf that may be overwrite in the
conf/local.conf with the desired path.
The recipes using the hardcoded /vsdk/demos are update to use the global
variable

Issue: ALB-2942

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>